[e2e] Stabilize standalone example tests
Generated-by: Codex CLI 0.144.5
